+++ skrub/_reporting/_table_report.py
@@ -18,6 +18,11 @@ class TableReport:
     ----------
     dataframe : pandas or polars DataFrame
         The dataframe to summarize.
+    n_rows : int, default=10
+        Maximum number of rows to show in the sample table. Half will be taken
+        from the beginning (head) of the dataframe and half from the end
+        (tail). Note this is only for display. Summary statistics, histograms
+        etc. are computed using the whole dataframe.
     order_by : str
         Column name to use for sorting. Other numerical columns will be plotted
         as function of the sorting column. Must be of numerical or datetime
@@ -30,17 +35,6 @@ class TableReport:
         mapping with the keys ``display_name`` (the name shown in the menu,
         e.g. ``"First 10 columns"``) and ``columns`` (a list of column names).
         See the end of the "Examples" section below for details.
-
-    Attributes
-    ----------
-    html : str
-        Report as an HTML page.
-    html_snippet : str
-        Report as an HTML snippet containing a single '<div>' element. Useful
-        to embed the report in an HTML page or displaying it in a Jupyter
-        notebook.
-    json : str
-        Report in JSON format.
 
     Notes
     -----
@@ -101,8 +95,20 @@ class TableReport:
     "b".
     """
 
-    def __init__(self, dataframe, order_by=None, title=None, column_filters=None):
-        self._summary_kwargs = {"order_by": order_by}
+    def __init__(
+        self,
+        dataframe,
+        n_rows=10,
+        order_by=None,
+        title=None,
+        column_filters=None,
+    ):
+        n_rows = max(1, n_rows)
+        self._summary_kwargs = {
+            "order_by": order_by,
+            "max_top_slice_size": -(n_rows // -2),
+            "max_bottom_slice_size": n_rows // 2,
+        }
         self.title = title
         self.column_filters = column_filters
         self.dataframe = dataframe
@@ -164,7 +170,7 @@ class TableReport:
         str :
             The JSON data.
         """
-        to_remove = ["dataframe", "head", "tail", "first_row_dict"]
+        to_remove = ["dataframe", "sample_table", "first_row_dict"]
         data = {
             k: v for k, v in self._summary_without_plots.items() if k not in to_remove
         }