+++ skrub/_reporting/js_tests/cypress/e2e/tabs.cy.js
@@ -2,6 +2,9 @@ describe('test tabbed interface navigation', () => {
     it('can select tab panel with mouse or keyboard', () => {
         cy.get('@report').find("[data-test='sample-tab']").as(
             'sampleTab').should('have.data', 'isSelected');
+        cy.get('@report').find("[data-test='summary-statistics-tab']")
+            .as(
+                'statsTab').should('not.have.data', 'isSelected');
         cy.get('@report').find("[data-test='summaries-tab']").as(
             'summariesTab').should('not.have.data', 'isSelected');
         cy.get('@report').find("[data-test='associations-tab']").as(
@@ -9,6 +12,9 @@ describe('test tabbed interface navigation', () => {
 
         cy.get('@report').find("[data-test='sample-panel']").as(
             'samplePanel').should('be.visible');
+        cy.get('@report').find("[data-test='summary-statistics-panel']")
+            .as(
+                'statsPanel').should('not.be.visible');
         cy.get('@report').find("[data-test='summaries-panel']").as(
             'summariesPanel').should('not.be.visible');
         cy.get('@report').find("[data-test='associations-panel']").as(
@@ -19,6 +25,9 @@ describe('test tabbed interface navigation', () => {
         cy.get('@summariesPanel').should('be.visible');
 
         cy.get('@summariesTab').should('have.focus');
+        // with modifier does nothing
+        cy.get('@summariesTab').type('{shift+rightArrow}');
+        cy.get('@summariesPanel').should('be.visible');
         cy.get('@summariesTab').type('{rightArrow}');
         cy.get('@summariesPanel').should('not.be.visible');
         cy.get('@associationsPanel').should('be.visible');
@@ -31,11 +40,11 @@ describe('test tabbed interface navigation', () => {
         cy.get('@sampleTab').should('have.focus');
         cy.get('@sampleTab').type('{rightArrow}');
         cy.get('@samplePanel').should('not.be.visible');
-        cy.get('@summariesPanel').should('be.visible');
+        cy.get('@statsPanel').should('be.visible');
 
-        cy.get('@summariesTab').should('have.focus');
-        cy.get('@summariesTab').type('{leftArrow}');
-        cy.get('@summariesPanel').should('not.be.visible');
+        cy.get('@statsTab').should('have.focus');
+        cy.get('@statsTab').type('{leftArrow}');
+        cy.get('@statsPanel').should('not.be.visible');
         cy.get('@samplePanel').should('be.visible');
     });
 });